Portability
This diff allows the Makefile to be a little better at guessing the host
platform, and adds some changes to allow the same source to compile on both
DietPi v6.30.0 (Raspbian Buster) RPi3B+ armv7l and Ubuntu 20.04 x86_64.

The included copy of libusb-1.0 has been removed so libusb-1.0-0-dev (etc.)
will need to be installed.

Some additional compatibility goo has been added to allow opencv 4.2.0 on
x86_64 to work just as well as opencv 3.2.0 on ARM.

ZWO SDK now includes a few other architectures; I'll leave it for folks with
armv8 systems to add the appropriate magic to the makefile.