start to remove absolute paths
Bash is clever enough to give us `$BASH_ARGV0` from which we can derive the
full path to the script, and thus the path to the top level of the currently
active allsky instance. If you decide to move the allsky directory... we just
recompute the path and everything (mostly) works normally.