Allow name of Pi to be specified for GUI install
Since I have multiple installations I needed a way to prevent the gui nameing all of them 'allsky'.
This change allows the hostname to be specified when installing the GUI. If no name is specified then the default 'allsky' is used.
Tested on a clean pi after installing allsky. Also re run the gui installer several times to confirm the hostname changes correctly.