Update saveImageNight.sh
Removed check for DARK_MODE since darkCapture.sh exits the script if DARK_MODE is on.
Moved "identity" check to top; there's no need to do anything else if the image is corrupt.
Perform error checking on "convert".
Allow user to specify thumbnail size and resize size.
Implement $IMG_PREFIX instead of hard-coding "liveview-".
Removed unnecessary double quotes on lftp line to make it easier to read.
The final image name is always the same now, so it's no longer "image-processed.jpg" if dark subtract is on, and "image.jpg" if it's off.  That was confusing.