Update timelapse.sh
Allows timelapses to be created for images no in ~/allsky/images.  For example, for observatory cameras.
Save "ln -s" output to temp file rather than outputting 1000+ lines to log file.  temp file is save so user can look at it, and a summary is output.
Perform error checking.
Decreased useless output from ffmpeg.
lftp puts to a temp name, then renames it to the correct name.  This is because timelapse files are large and if the web server tries to read the file as it's being uploaded, it will get an error.