Create videos/thumbnails
This changes goes with the changes to "allsky-website" that automatically creates video thumbnails.
Once people grab the "allsky-website" changes they'll need to create the vidoes/thumnails directory.  Should we move the "Create" and "Fixing" code block to a different shell script that users can invoke, and have install.sh also invoke it?  That should make it easier for non-technical users to create the directory.  The "mkdir" will need to change to "mkdir -p".