Catch aborts by capture program
When the "capture" program aborts/crashes (versus gracefully exits), the exit code is (usually?) greater than 100.  When that happens, for example due to a bug, there's no reason to let the  allsky service restart since the program will crash again, creating an endless cycle.  allsky.sh notices when the capture program aborts and displays an "Error" image directing the user to look in the log file and fix the problem.  It then stops the service.