Multiple bug fixes and improvements
* Disable alwaysshowadvanced for default

It was accidentally left as "1" from my Pi.

* Create videos/thumbnails

This changes goes with the changes to "allsky-website" that automatically creates video thumbnails.
Once people grab the "allsky-website" changes they'll need to create the vidoes/thumnails directory.  Should we move the "Create" and "Fixing" code block to a different shell script that users can invoke, and have install.sh also invoke it?  That should make it easier for non-technical users to create the directory.  The "mkdir" will need to change to "mkdir -p".

* Update README.md

Fixes #442 
and adds some additional clarifications.

* Fixed check for LF and CR

The "free(line);" was causing the problem.  I must be misreading the getline() man page where it says to free() the line...

* Catch aborts by capture program

When the "capture" program aborts/crashes (versus gracefully exits), the exit code is (usually?) greater than 100.  When that happens, for example due to a bug, there's no reason to let the  allsky service restart since the program will crash again, creating an endless cycle.  allsky.sh notices when the capture program aborts and displays an "Error" image directing the user to look in the log file and fix the problem.  It then stops the service.

* Fix to day-to-night transition

Fixes #443