unbreak ROI selection (#474)
* unbreak ROI selection

New exposure mode did not correctly set the capture ROI if the image size
was manually specified. It was calling SetROI with a 0x0 window which would
cause subsequent captures to fail.

* exit if ROI set fails

Not exiting with a high error code since it might be a transitory error,
or something the user can fix quickly based on the log message.