Autodetect minimum exposure for buffer clear
Instead of hardcoding something, ask the camera what its minimum exposure
time is, then use it. Fall back to a generous 100ms if the query fails...

... which I can't see happening since we might already query some of the
capabilities at startup. I don't see any reason to abort the program if
this query fails, just log a debug message for the curious.