add getopt to keogram
This is a breaking change and will require `endOfNight.sh` to be updated to
use the new argument format. On the upside, the help screen is prettier, main()
is simpler, and it allows me to do things like

`./keogram -d /tmp/20210919/ -e jpg -o /tmp/k.jpg --font-color '#00ff00' --font-name triplex`

and not be spammed by tons of logs. Of course, if you like that sort of thing,
you are welcome to add `-v` or `--verbose`.