add getopt to startrails
This is a breaking change and will require `endOfNight.sh` to be updated to
use the new argument format. On the upside, the help screen is prettier, main()
is simpler, and it allows me to do things like

	`./startrails -b 0.3 -d /tmp/20210919/ -e jpg -o /tmp/s.jpg`

and not be spammed by tons of logs. Of course, if you like that sort of thing,
you are welcome to add `-v` or `--verbose`.

New feature: summary mode. Just scan the directory without rendering an image:
	`./startrails -s -d /tmp/20210919 -e jpg`