Another way OpenCV could explode.
What is it with my system, finding all of these broken images?!

Anyway, somehow I've got a few monochrome images scattered through
my directory of RGB, which of course makes OpenCV get its knickers
in a knot. To mitigate this, I take the first image of appropriate
dimension, and use its channel count as authoritative. Any images
not matching that will be converted.