more improvements
allow font color to be specified in opencv order as a string of 3 ints
representing blue, green and red channels, eg. "255 128 64" as well as
in the usual rgb hex format, eg. "c0ffee" or "#3333cc".

and add a "-p"/"--parse-filename" flag to derive the image time from the
filename rather than its time on disk (which may not be preserved across
copies. Fixes #503