Silence aux tools (#493)
* add getopt to keogram

This is a breaking change and will require `endOfNight.sh` to be updated to
use the new argument format. On the upside, the help screen is prettier, main()
is simpler, and it allows me to do things like

`./keogram -d /tmp/20210919/ -e jpg -o /tmp/k.jpg --font-color '#00ff00' --font-name triplex`

and not be spammed by tons of logs. Of course, if you like that sort of thing,
you are welcome to add `-v` or `--verbose`.

* add getopt to startrails

This is a breaking change and will require `endOfNight.sh` to be updated to
use the new argument format. On the upside, the help screen is prettier, main()
is simpler, and it allows me to do things like

	`./startrails -b 0.3 -d /tmp/20210919/ -e jpg -o /tmp/s.jpg`

and not be spammed by tons of logs. Of course, if you like that sort of thing,
you are welcome to add `-v` or `--verbose`.

New feature: summary mode. Just scan the directory without rendering an image:
	`./startrails -s -d /tmp/20210919 -e jpg`

* make endOfNight.sh know about the new arguments to keogram and startrails

Since keogram and startrails no longer gratuitously emit a log line per file
processed, it is no longer necessary to create additional log files.

Grrrr, vscode fixed all the tabs 'n' spaces.

* fix directory long_opt name

* forgot to spell out 'rotate'