unify multiprocessor flags
now both of these tools use -Q / --max-threads to control the number of worker
threads, and -q / --nice-level to set the scheduler priority.

Q is vaguely mnemonic for queue (job scheduling) and it's not likely that you
will want to change the settings without a good reason. Use as much CPU as
possible, without dragging the whole system down, right?