generateForDay.sh: consolidate code
Keogram, startrails, and timelapse all had almost identical code to create them and to upload them - they all used the same logic.
This commit moves that logic into "generate" and "upload" functions, resulting in a smaller file with no duplicate code.

NOTE: when looking at the changes in this commit, Git does a poor job - it compares the two new functions to the old keogram code, and there are a bunch of differences.  It doesn't seem to notice that the two functions are new.  If you look at the startrails and timelapse differences it'll be much clearer.