darkSubtract.sh: don't force -type TrueColor
The "convert" command is smart enough to determine what bit depth the output file should be based on the input file and the dark file, so we don't need to force TrueColor (24 bit).  This causes problems because daytime mono images are 8-bit (no dark subtraction is done) but nighttime images are 24-bit.