darkCapture.sh: accept different $EXTENSION
Don't assume the extension of the dark frame is the same as the light frame.  It the future the dark will likely by .png and the light .jpg.
Work with $CURRENT_IMAGE passed to us (in the future), or not passed to us (present).
Use $DARKS_DIR instead of $ALLSKY_DARKS, since in the future when darks are supported per-camera, DARKS_DIR will point to the camera-specific directory.
Also added {} around some variables.
Added/changed some comments for clarity.