capture_RPiHQ.cpp: compiler warning
The "extra text" file hasn't yet been fully implemented in the RPiHQ capture program, but variables to activate it are there.  This causes compiler messages about "variable set but not used".  This is not an issue, but some users have mentioned it, thinking it's an error.
I added a dummy line to keep the compiler quiet.  This line will be removed when the "extra text" file is implemented.

Fixes #833.