saveImage.sh: improve efficiency
Instead of moving the image from A to B, then copying it back to A, simply copy from A to B, then rename A.
When resizing the image for upload, don't first copy and then run "convert" on it (which will overwrite the file).  Instead, just run "convert" to create the resized file using a unique file name to lessen the chance of it being overwritten by another process.
When not resizing the image, use the final image in the $ALLSKY_IMAGES/<date> directory which has a unique name, as the file being uploaded.  This will also lessen the chance of the file being overwritten (was it was called "image.jpg") as it's being uploaded (which could result in a temporary file being left on the web server).