postData.sh: fix inaccuracy
Use the user Angle instead of "civil" twilight which assumes an angle of -6.
Copy the data.json file to the website on the Pi, if it exists.  This is needed so the local website knows when daylight is and can display the aurora forecast picture if daytime images are not being taken.
Warn the user if their angle results in an unknown sunset time from sunwait.
Warn the user if there is no local website and no remote host to copy the file to.
Improve comments.