darkSubtract.sh: remove leading zero's
When converting a string to a number, if the string begins with a 0, it's treated as an octal number.  Unfortunately, if the THIS_TEMPERATURE variable is "09" that gives an error when converted to integer since 9 is not a legal octal number.
I couldn't find anyway around this.
I will also change capture.cpp to not add a leading 0, but it's possible a user script stores numbers with a leading 0 in tmp/temperature.txt.