install.sh: All updating over prior version
Take into account that the user may already have the Allsky Website installed, so save it, then install the new version, and restore saved videos, keograms, and startrails as well as any associated thumbnails.  Also restore configuration files.
If the user is installing Allsky Website for the first time, set some configuration variables based on the Allsky configuration.

The code dealing with configuration files was moved to a function "modify_configuration_variables" in case we want to call it on "--update".