install.sh: change XX_ON_PI_XX
functions.php needs to know if it's on a Pi so it can check the "XX_ALLSKY_CONFIG_XX" placeholder and if not updated, warn the user.
On remote servers we should NOT check the placeholder.  The "XX_ON_PI_XX" placeholder is set "true" during installation on a Pi which tells functions.php it's running on a Pi.

There is no longer a need to replace only the 1st instance of XX_ALLSKY_CONFIG_XX since there is now only one instance.