cleanup: makeSunwaitWarningFree
Please excuse this dirty solution. Multiple calls can lead to the flag "-w" being inserted multiple times - but this has no effect on the function
But the main advantage is that sunwait is now free of warnings. This avoids concerned inquiries from users and we can now concentrate on our own warnings and errors