allsky.sh: improvements 2
Handle exit codes 98 and 99 from the capture programs (which will be added via a separate PR).
Add initial code for trying to determine if the service is restarting, versus stopping.  On a restart we can display a message "Allsky software is restarting" and skip the "Allsky software is not running" message that appears when the service stops.  I'm pretty sure there's a way with systemctl that allows it to send a different signal on stop versus restart.  allsky.sh and the capture programs could then check which signal was sent.