config.sh.repo: WEBUI_DATA_FILES
Replace the FAN_DATA_FILE with WEBUI_DATA_FILES which is a generic way to add data, progress bars, and/or buttons to the WebUI "System" page.  The old FAN_DATA_FILE variable only supported adding fan data.  The new mechanism is much more generic and powerful.
There is a related PR in the allsky-portal package to implement the code changes.