mode_mean.cpp: change multipliers
The old single multiplier of 2 * mean_threshold could result in "slow forward" when the mean was pretty far away from the target, e.g., mean == 0 and target == 0.2.  Therefore, lower the multiplier and add a second one with an even lower multiplier that increases the ExposureChange less.
Also, use the difference is the actual mean and the target mean, not the forecasted or expected new mean (which aren't always accurate).