Update timelapse.sh
timelapse.sh now just creates a mini-timelapse when asked to.  It no longer cares if there are "enough" images. This means we don't need to pass any numbers to it, just the file that contains the images to be used in the timelapse.
We also no longer update the file that contains images - we let the invoker do that.
Both these things significantly simplified timelapse.sh.