Update saveImage.sh
The mini-timelapse code didn't work very well - the file that contains the list of images to include wasn't always updated, so most of the mini-timelapses were missed.
Since timelapse.sh no longer checks how many images there are or updates the file with image names, we do that here.  We were trying to count the images before but it often wasn't in sync with the file contents.
Now the count is not stored - it's always gathered by the number of entries in the file.