functions.php: fix
Turns out, displaying the default for empty values isn't good.  The settings file may have "" but the user sees the default.  The Allsky Settings page should always display what's in (or not in) the settings file.  Will do a different PR to not allow blank settings if they are required.