Update editor.php
save_file.php can contain a single line of output which should start with "S\t" for success or "E\t" for error.
However, it calls updateFile() which may produce debug output.  In that case we need to look through every line of output for the status (S or E) and "other" lines.  For the "other" lines, send them to console.log().