check_allsky.sh: more intelligent image check
When determining if an image size is "too large" and may cause a problem with timelapse video creation, use CROP_IMAGE, IMG_RESIZE, and the WebUI size (in that order), THEN use the sensor size if none of the others are set.  This should give us the actual size of the image.