Only prompt for new hostname if it's currently "raspberrypi"
I don't think we really care what the name of the Pi is, although the documentation assumes "allsky".  For machines that still use the default "rapberrypi", prompt for a new name with "allsky" as the default.
We don't prompt for other names since they imply the user already change the name.