Update timelapse.sh: fix bug with concurrent mini timelapse
If timelapse.sh removes the lock file it's possible saveImage.sh (which calls timelapse.sh) will be uploading a mini-timelapse when another timelapse.sh is called.  With no lock file, it will start creating a new timelapse which breaks the upload that's also in progress.
We get around this by using timelapse.sh's parent PID (saveImage.sh) and let IT remove the lock file when done uploading.