Update allskySettings.php: bunch of changes dealing with booleans
getVariableOrDefault() has to return booleans as strings (e.g., "true") because when returning as booleans (e.g., true), false values are returned as "".

However, we want booleans saved in the settings file so we have to convert from strings to booleans.  What a pain.

A bunch of debug statements were left in case there's a problem in the future.  Once we know this is rock solid many of the debug statements can be deleted.