Update timelapse.sh: Allow non-allsky files to be used
Likely a very rare usage, but I needed it!
I have a bunch of files from a security camera and wanted to create a timelapse of them.  Their names do NOT begin with "image-" so timelapse would always fail.
The "--filename file" option allows any set of files to be used, as long as their names all begin with "file" and they have the same extension as Allsky's file does.