Update allsky.sh: Ignore signals so we continue after capture ends
When the allsky service stops allsky.sh and all its children (including the capture program) get the signal.  They all die and the code in allsky.sh after the call to the capture program does not run.

allsky.sh needs to ignore the signals.  The capture program will still get the signal and die.