Update allskySettings.php: Fix initial "+" being deleted in lat/long.
This fix is quite complex. JSON_NUMERIC_CHECK has the nasty side effect of removing leading "+" in strings that look like numbers, as can be used in latitude and longitude.
In order to not use JSON_NUMERIC_CHECK all numbers must be converted to either PHP "integer" or "float" so json_encode() knows how to save them.  This requires knowing what type each setting is.