Update functions.php: calculate more accurate message
When autoexposure is ON we really don't know how long the exposure time will be so we don't know the time from one picture start to the next.
Because daytime pictures are usually less than 1 second, assume a 1 second exposure.  Nighttime exposure are usually 10 or more seconds, so assume a 10-second exposure.